fix: refactor code for "add instrument name"
- remove "alias" field from "InstrumentName" model;
- move "publish_name" to a separate .py file
- use "csrf_protect" decorator for "publish_name"
- extract the common code in "mansonryView.html" and "stdView.html"

Refs: #163
