+++ web-app/django/VIM/apps/instruments/templates/instruments/includes/instrumentContainer.html
@@ -1,0 +1,28 @@
+{% load static %}
+
+<div class="instrument-img-container">
+  <img src="{% static instrument.thumbnail.url %}" class="card-img-top rounded p-2 img-fluid instrument-img" alt="instrument thumbnail" onerror="this.onerror=null;this.src='{% static "instruments/images/no-image.svg" %}';" />
+  <div class="middle-button-group">
+    {% if user.is_authenticated %}
+      <button type="button"
+              class="btn"
+              data-bs-toggle="modal"
+              data-bs-target="#addNameModal"
+              data-instrument-name="{{ instrument.instrumentname_set.first.name|title }}"
+              data-instrument-wikidata-id="{{ instrument.wikidata_id }}">
+        Add instrument name
+      </button>
+      <button type="button"
+              class="btn"
+              data-bs-toggle="modal"
+              data-bs-target="#uploadImagesModal"
+              data-instrument-name="{{ instrument.instrumentname_set.first.name|title }}"
+              data-instrument-wikidata-id="{{ instrument.wikidata_id }}">
+        Upload new images
+      </button>
+    {% endif %}
+    <a class="btn"
+       href="https://www.wikidata.org/wiki/{{ instrument.wikidata_id }}"
+       target="_blank">View on Wikidata</a>
+  </div>
+</div>
