+++ web-app/django/VIM/apps/instruments/templates/instruments/includes/masonryView.html
@@ -6,37 +6,11 @@
     <div class="col">
       <div class="text-decoration-none">
         <div class="card mb-3">
-          <div class="instrument-img-container">
-            <img src="{% static instrument.thumbnail.url %}" class="card-img-top rounded p-2 instrument-img" alt="instrument thumbnail" onerror="this.onerror=null;this.src='{% static "instruments/images/no-image.svg" %}';" />
-            <div class="middle-button-group">
-              {% if user.is_authenticated %}
-                <button type="button"
-                        class="btn"
-                        data-bs-toggle="modal"
-                        data-bs-target="#addNameModal"
-                        data-instrument-name="{{ instrument.instrumentname_set.first.name|title }}"
-                        data-instrument-wikidata-id="{{ instrument.wikidata_id }}">
-                  Add instrument name
-                </button>
-                <button type="button"
-                        class="btn"
-                        data-bs-toggle="modal"
-                        data-bs-target="#uploadImagesModal"
-                        data-instrument-name="{{ instrument.instrumentname_set.first.name|title }}"
-                        data-instrument-wikidata-id="{{ instrument.wikidata_id }}">
-                  Upload new images
-                </button>
-              {% endif %}
-              <a class="btn"
-                 href="https://www.wikidata.org/wiki/{{ instrument.wikidata_id }}"
-                 target="_blank">View on Wikidata</a>
-            </div>
-          </div>
+          {% include "instruments/includes/instrumentContainer.html" %}
+
           <div class="card-body pb-0 pt-0">
             <p class="card-title text-center notranslate ">
-              {% for instrumentname in instrument.instrumentname_set.all %}
-                {{ instrumentname.name|title }}
-              {% endfor %}
+              {{ instrument.instrumentname_set.all.first.name|title }}
             </p>
           </div>
         </div>
